requirements.txt
structlog
pyyaml
python-dotenv
boto3

langchain
langchain-text-splitters
langchain-google-genai
langchain-openai
openai
langchain-groq
langchain-qdrant

PyMuPDF
docx2txt
python-pptx

-e .